Rename Validator to Parser
- Changed Validator class to Parser
- Updated all references in Orion.java
- Moved Parser.java to correct package

Part of ongoing refactoring for A-MoreOOP task.
