+++ src/main/java/Orion.java
@@ -4,7 +4,7 @@ import task.DeadlineDetails;
 import task.EventDetails;
 import task.Task;
 import taskList.TaskList;
-import validator.Validator;
+import parser.Parser;
 import orionExceptions.*;
 import commands.Command;
 
@@ -21,7 +21,7 @@ public class Orion {
         }
     }
 
-    public static Validator validator = new Validator();
+    public static Parser parser = new Parser();
     private static final String HORIZONTAL_LINE = "────────────────────────────────────────";
 
     public static void main(String[] args) {
@@ -73,7 +73,7 @@ public class Orion {
     private static void handleList(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                validator.validateListCommand(parts);
+                parser.validateListCommand(parts);
                 manager.printTasks();
             } catch (OrionException e) {
                 System.out.println(e.getMessage());
@@ -84,7 +84,7 @@ public class Orion {
     private static void handleEvent(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                EventDetails details = validator.validateEventCommand(parts);
+                EventDetails details = parser.validateEventCommand(parts);
                 Task temp = manager.addEvent(details);
                 int taskLen = manager.getSize();
                 System.out.println("     Got it. I've added this task:\n" + temp + "\n" + "Now you have " + taskLen + " tasks in the list");
@@ -98,7 +98,7 @@ public class Orion {
     private static void handleDeadline(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                DeadlineDetails deadlineDetails = validator.validateDeadlineCommand(parts);
+                DeadlineDetails deadlineDetails = parser.validateDeadlineCommand(parts);
                 Task newDeadline = manager.addDeadline(deadlineDetails);
                 System.out.println("Got it. I've added this task:");
                 System.out.println("  " + newDeadline);
@@ -111,7 +111,7 @@ public class Orion {
     private static void handleTodo(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                String description = validator.validateTodoCommand(parts);
+                String description = parser.validateTodoCommand(parts);
 
                 Task temp = manager.addTodo(description);
                 int taskLen = manager.getSize();
@@ -126,7 +126,7 @@ public class Orion {
     private static void handleMark(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                int index = validator.validateMarkAndUnMarkCommand(parts, manager);
+                int index = parser.validateMarkAndUnMarkCommand(parts, manager);
                 Task temp = manager.markAsDone(index);
                 System.out.println("Marked task as done: " + temp);
             } catch (OrionException e) {
@@ -138,7 +138,7 @@ public class Orion {
     private static void handleUnmark(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                int index = validator.validateMarkAndUnMarkCommand(parts, manager);
+                int index = parser.validateMarkAndUnMarkCommand(parts, manager);
                 Task temp = manager.unmarkAsDone(index);
                 System.out.println("Unmarked task: " + temp);
             } catch (OrionException e) {
@@ -150,7 +150,7 @@ public class Orion {
     private static void handleDelete(String[] parts) {
         executeWithFormatting(() -> {
             try {
-                int index = validator.validateDeleteCommand(parts, manager);
+                int index = parser.validateDeleteCommand(parts, manager);
                 Task deletedTask = manager.deleteTask(index);
                 System.out.println("Noted. I've removed this task:");
                 System.out.println("  " + deletedTask);
